Split BindingsTests into multiple files
- Move user setup to before tests are run
- Move shared User objects to the User helper class
- Move APIResponse Body Parser to RequestBodyHelper
