+++ e2e-tests/build.gradle
@@ -78,7 +78,7 @@ dependencies {
   testImplementation 'org.junit.jupiter:junit-jupiter-params:6.0.1'
 
   testRuntimeOnly 'org.junit.platform:junit-platform-engine:6.0.1'
-  testRuntimeOnly 'org.junit.platform:junit-platform-launcher:6.0.1'
+  testImplementation 'org.junit.platform:junit-platform-launcher:6.0.1'
 }
 
 tasks.register('buildAllProcedureJars') {
