+++ e2e-tests/src/test/resources/META-INF/services/org.junit.platform.launcher.LauncherSessionListener
@@ -1,0 +1,1 @@
+gov.nasa.jpl.aerie.e2e.utils.UserSetup
